fix(core): funky type issues with the main export exporting multiple things
