+++ packages/core/package.json
@@ -17,6 +17,10 @@
       "require": "./dist/lib/index.cjs",
       "import": "./dist/lib/index.js"
     },
+    "./types": {
+      "require": "./dist/types.d.cts",
+      "import": "./dist/types.d.ts"
+    },
     "./package.json": "./package.json"
   },
   "main": "dist/index.cjs",
