+++ packages/core/src/index.ts
@@ -1,3 +1,4 @@
+import type { ConfigOptions } from './types.js';
 import type { Har } from 'har-format';
 import type Operation from 'oas/operation';
 import type { HttpMethods, OASDocument } from 'oas/rmoas.types';
@@ -8,28 +9,6 @@ import Oas from 'oas';
 
 import FetchError from './errors/fetchError.js';
 import { parseResponse, prepareAuth, prepareParams, prepareServer } from './lib/index.js';
-
-export interface ConfigOptions {
-  /**
-   * Override the default `fetch` request timeout of 30 seconds. This number should be represented
-   * in milliseconds.
-   */
-  timeout?: number;
-}
-
-export interface FetchResponse<HTTPStatus, Data> {
-  data: Data;
-  headers: Headers;
-  res: Response;
-  status: HTTPStatus;
-}
-
-// https://stackoverflow.com/a/39495173
-type Enumerate<N extends number, Acc extends number[] = []> = Acc['length'] extends N
-  ? Acc[number]
-  : Enumerate<N, [...Acc, Acc['length']]>;
-
-export type HTTPMethodRange<F extends number, T extends number> = Exclude<Enumerate<T>, Enumerate<F>>;
 
 export default class APICore {
   spec!: Oas;
