+++ packages/test-utils/sdks/operationid-quirks/src/index.ts
@@ -1,4 +1,4 @@
-import type { ConfigOptions, FetchResponse } from '@readme/api-core';
+import type { ConfigOptions, FetchResponse } from '@readme/api-core/types';
 import APICore from '@readme/api-core';
 import definition from '@api/test-utils/definitions/operationid-quirks.json';
